% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yz_motifenrich.R
\name{addmotifLogo}
\alias{addmotifLogo}
\title{make a html file}
\usage{
addmotifLogo(motifEnrDT, addHTML = TRUE, motifcolumname = "motifrowname")
}
\arguments{
\item{motifEnrDT}{is a dataframe include motifTFname,motifrowname,similarity fuction:compMotif}

\item{motifcolumname}{you can chose which to add the logo ,you must build a index}
}
\description{
make a html file
}
