% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atac.R
\name{atac_peakanno}
\alias{atac_peakanno}
\title{ATACseq peak annotation}
\usage{
atac_peakanno(
  peaks,
  txdb,
  rnaseqtxt,
  tss_region = 3000,
  prefix,
  tfname = "WRKY"
)
}
\arguments{
\item{peaks}{the peak call file from the fseq2 of GRanges object}

\item{txdb}{you interested species annotation genome file of TxDb object}

\item{tss_region}{the number of base pair from tss ,default is 3000}

\item{prefix}{the name for save your result}

\item{tfname}{the TF family name can be plot with RNA-seq and ATAC-seq,default is WRKY}
}
\description{
ATACseq peak annotation
}
