% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zfj_motif_202108.R
\name{autoseed_motif_disc}
\alias{autoseed_motif_disc}
\title{Motif Discovery with Autoseed}
\usage{
autoseed_motif_disc(
  sig_file,
  outdir = NULL,
  bk_file = "/var/www/html/selex/data/sequence/Empty.seq",
  local_max_cutoff = 40
)
}
\arguments{
\item{sig_file}{a file containing the reads for motif discovery, can be either a fastq, fasta, or plain text file}

\item{outdir}{the path to store output data (seeds and svg visualizations)}

\item{bk_file}{the background file for motif discovery, leave blank to use random sequences}

\item{local_max_cutoff}{the sensitivity of motif discovery (it tells at least how many hits of the seed consensus should exist in the sig_file, in order to generate a PFM output for this seed )}
}
\value{
a list of pfms discovered from the sig_file, can use ggseqlogo for further visualization
}
\description{
Motif Discovery with Autoseed
}
\examples{
sig_file="/var/www/html/selex/data/sequence/random40ND8.44.ATI4.seedling.root.20101n.1..30n4u_sig.seq"
result=autoseed_motif_disc(sig_file,"/wrk/zhu/test/autoseed_test",local_max_cutoff = 20)
}
