% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yz_motifenrich.R
\name{compMotif}
\alias{compMotif}
\title{find the de novo motif with known TFs motif}
\usage{
compMotif(uniquedir, outdir)
}
\arguments{
\item{uniquedir}{the dir include the unique motif matrix file}

\item{outdir}{the motif logos}
}
\description{
find the de novo motif with known TFs motif
}
