% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yz_motifenrich.R
\name{copMotif}
\alias{copMotif}
\title{chose the unique motif matrix}
\usage{
copMotif(inputdir, outdir)
}
\arguments{
\item{outdir}{the dir remove the repeat motif matrix}

\item{splitmotifdir}{the dir include the split motif matrix file}
}
\description{
chose the unique motif matrix
}
