% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif_html.R
\name{.addmotifLogorow}
\alias{.addmotifLogorow}
\title{make html index}
\usage{
.addmotifLogorow(
  motifEnrDT,
  addHTML = TRUE,
  motifcolumname = "motifTF_name",
  motifsimi = "motifTF_similarity",
  name_m = pcm["name"]
)
}
\arguments{
\item{motifEnrDT}{a dataframe include motifTF_name,motifTF_family,motifTF_similarity,and so on}

\item{addHTML}{add picture index ,default is TRUE}

\item{motifcolumname}{the column name of motif in motifEnrDT}

\item{motifsimi}{the column name of similarity in motifEnrDT}

\item{name_m}{the motif name}
}
\value{

}
\description{
make html index
}
\seealso{
\code{\link[=motif_database_html]{motif_database_html()}}
}
