% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif_html.R
\name{.da_sim}
\alias{.da_sim}
\title{find motif information in database}
\usage{
.da_sim(x_n, pcm, pwm_library, motifdatabase, name_, ge)
}
\arguments{
\item{x_n}{a data.frame include tf_name,similarity,and the location in database}

\item{pcm}{you give consensus or universalmotif format motif}

\item{pwm_library}{a PWMartrixList of database motif}

\item{motifdatabase}{a universalmotif format list of database motif}

\item{name_}{the number of location}

\item{ge}{the text file of gene_id and TF_name,famliy or other informations}
}
\value{

}
\description{
find motif information in database
}
\seealso{
\code{\link[=motif_database_html]{motif_database_html()}}
}
