% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zfj_motif_202108.R
\name{.length_to_40bp}
\alias{.length_to_40bp}
\title{adjust sequence length to 40bp in order to use autoseed for motif discovery}
\usage{
.length_to_40bp(in_file, out_file, segment_x3 = TRUE)
}
\arguments{
\item{segment_x3}{}
}
\value{

}
\description{
adjust sequence length to 40bp in order to use autoseed for motif discovery
}
