% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zfj_motif_202108.R
\name{.pfm_trim}
\alias{.pfm_trim}
\title{trim low IC bases flanking a pfm}
\usage{
.pfm_trim(pfm, include_flank_cutoff = 2)
}
\arguments{
\item{include_flank_cutoff}{}
}
\value{

}
\description{
trim low IC bases flanking a pfm
}
