% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{heatmap_kmer}
\alias{heatmap_kmer}
\title{you can make a heatmap to see your fasta file kmer}
\usage{
heatmap_kmer(sampleinfor, outpath)
}
\arguments{
\item{sampleinfor}{a tibble or datafram include the information of batch well tissue ligand length and file dress}

\item{outpath}{to save the results}
}
\value{
different ligand length heatmap png and pdf
}
\description{
you can make a heatmap to see your fasta file kmer
}
