% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zfj_misc_utilities_202111.R
\name{is_from_adaptor}
\alias{is_from_adaptor}
\title{Check if a sequence is from Truseq adaptor}
\usage{
is_from_adaptor(
  test_seq = "ACGTGTGCTCTTC",
 
    lib_seq = "AATGATACGGCGACCACCGAGATCTACACTCTTTCCCTACACGACGCTCTTCCGATCTnnnnnnnnnnnnnnnnnnnnnnnnnnnnnnnnAGATCGGAAGAGCACACGTCTGAACTCCAGTCACnnnnnnnnATCTCGTATGCCGTCTTCTGCTTG"
)
}
\arguments{
\item{test_seq}{the sequence or motif consensus to compare with Truseq ligand}

\item{lib_seq}{the sequence of full length Truseq ligand}
}
\value{
the result of alignment of test_seq, lib_seq, and the reverse complement of test_seq
}
\description{
Check if a sequence is from Truseq adaptor
}
