% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yz_motif_enrich_fjEdit202111.R
\name{motif_counts_}
\alias{motif_counts_}
\title{motif enrichment analysis for a target library}
\usage{
motif_counts_(seqs, pwmlist)
}
\arguments{
\item{seqs}{a DNA sequence file or a character vector of sequences}

\item{pwmlist}{motifs of PWMartixList format}
}
\value{

}
\description{
count motif occurances in a target file (or character vector) of sequences, the frequency is normalized to a background library with the same 1-nt frequency but got randomized
}
\examples{
pwmlist="/wrk/zhu/test/motif_curation/curated_30N_motif_pwmlist_20211014.rds" \%>\% readRDS()
seqs="/wrk/wenchenjin/work/Chenjin1_1__ATI_ara.cell_1/pre_processed/cleaned/B8_20_Mock__root__conc_0nM__time_2h__30n.gz"
motif_count <- motif_counts_(seqs,pwmlist)
}
