% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif_html.R
\name{motif_database_html}
\alias{motif_database_html}
\title{identify gave consensus or universalmotif motif to get TF html file}
\usage{
motif_database_html(pcm, n = 5, by = "pcm")
}
\arguments{
\item{pcm}{a motif of universalmotif format (PWM PCM or PPM)}

\item{n}{the result number you want to get,defualt is 5}

\item{by}{if consensus,use "by="consensus"".if universalmotif,use "by="pcm""}
}
\value{

}
\description{
identify gave consensus or universalmotif motif to get TF html file
}
