% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif_from_JP.R
\name{motif_from_JP}
\alias{motif_from_JP}
\title{make a pcm to the most similar known TFs motif name}
\usage{
motif_from_JP(pcm, species = "Arabidopsis thaliana", n = 1)
}
\arguments{
\item{pcm}{include your motif pcm}

\item{species}{your interested species(the JASPAR included),default is Arabidopsis thaliana}

\item{n}{the number of similar motif name,default is 1}
}
\description{
make a pcm to the most similar known TFs motif name
}
