% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yz_motif_enrich_fjEdit202111.R
\name{motif_get_name}
\alias{motif_get_name}
\title{match your motif to the most similar known motif form database}
\usage{
motif_get_name(PFMatrix)
}
\arguments{
\item{PFMatrix}{}
}
\value{

}
\description{
this return an TFBSTools format motif ,it's ID includes the name,family and similarty of known motif
}
\examples{
pwmlist="/wrk/zhu/test/motif_curation/curated_30N_motif_pwmlist_20211014.rds" \%>\% readRDS()
motif_get_name(pwmlist[[1]])
}
\seealso{
\code{\link[=motif_database_html]{motif_database_html()}},to see which known motif include
}
