% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yz_motifenrich.R
\name{plotEnrich}
\alias{plotEnrich}
\title{drow a fold changes and -log10 picture}
\usage{
plotEnrich(sequencefile, prefix, backgroundseq = NULL)
}
\arguments{
\item{sequencefile}{your raw data sequence file must fasta file}

\item{prefix}{name of the output}

\item{backgroundseq}{background sequence also fasta file,and if NULL the file will  automatic generation}
}
\description{
drow a fold changes and -log10 picture
}
