% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yz_motifenrich.R
\name{splitFile}
\alias{splitFile}
\title{split the total matrix file}
\usage{
splitFile(inputfile, output)
}
\arguments{
\item{inputfile}{the totall motif matrix txt file}

\item{output}{the split results}
}
\description{
split the total matrix file
}
